requirements.txt
Pillow==10.3.0
pycryptodome==3.20.0
tk==0.1.0
